fixing issue #33 - vypisovat zaznam ulozenych vysledku hry
git-svn-id: https://www.literak.cz/svn/lelisoft/LeliMath@157 da24d14a-7dfa-43c6-8752-2aa101e44d5d

+++ app/src/main/java/lelisoft/com/lelimath/activities/BaseGameActivity.java
@@ -36,7 +36,7 @@ public class BaseGameActivity extends LeliBaseActivity {
 
     protected void storeFormulaRecord(FormulaRecord record) {
         try {
-            Dao<FormulaRecord, Long> dao = getHelper().getFormulaRecordDao();
+            Dao<FormulaRecord, Integer> dao = getHelper().getFormulaRecordDao();
             int result = dao.create(record);
             if (result == 1) {
                 log.debug("Saved FormulaRecord");
